Update to 1.17 with Matrix stuffs

+++ src/com/projectkorra/rpg/worldevent/RPGWorldEvent.java
@@ -81,7 +81,7 @@ public interface RPGWorldEvent {
 	 * @return numerical modifier for bending abilities of the WorldEvent's
 	 *         element
 	 */
-	public double getModifier();
+	public long getModifier();
 	
 	/**
 	 * Returns how long the event lasts
